Rename DateTimeBranchOperator to BranchDateTimeOperator (#14720)
* Update names and paths to be consistent with other Branch operators

This included renaming the operator to match the '^Branch.*' pattern, like other operators do, as well as updating file and import paths to remove deprecated files (for example, changing the deprecated dummy_operator to dummy).

* Add missing reference to operator guide
