+++ airflow/example_dags/example_branch_datetime_operator.py
@@ -23,8 +23,8 @@ targets.
 import datetime
 
 from airflow import DAG
-from airflow.operators.datetime_branch import DateTimeBranchOperator
-from airflow.operators.dummy_operator import DummyOperator
+from airflow.operators.datetime import BranchDateTimeOperator
+from airflow.operators.dummy import DummyOperator
 from airflow.utils.dates import days_ago
 
 args = {
@@ -32,18 +32,18 @@ args = {
 }
 
 dag = DAG(
-    dag_id="example_datetime_branch_operator",
+    dag_id="example_branch_datetime_operator",
     start_date=days_ago(2),
     default_args=args,
     tags=["example"],
     schedule_interval="@daily",
 )
 
-# [START howto_operator_datetime_branch]
+# [START howto_branch_datetime_operator]
 dummy_task_1 = DummyOperator(task_id='date_in_range', dag=dag)
 dummy_task_2 = DummyOperator(task_id='date_outside_range', dag=dag)
 
-cond1 = DateTimeBranchOperator(
+cond1 = BranchDateTimeOperator(
     task_id='datetime_branch',
     follow_task_ids_if_true=['date_in_range'],
     follow_task_ids_if_false=['date_outside_range'],
@@ -54,21 +54,21 @@ cond1 = DateTimeBranchOperator(
 
 # Run dummy_task_1 if cond1 executes between 2020-10-10 14:00:00 and 2020-10-10 15:00:00
 cond1 >> [dummy_task_1, dummy_task_2]
-# [END howto_operator_datetime_branch]
+# [END howto_branch_datetime_operator]
 
 
 dag = DAG(
-    dag_id="example_datetime_branch_operator_2",
+    dag_id="example_branch_datetime_operator_2",
     start_date=days_ago(2),
     default_args=args,
     tags=["example"],
     schedule_interval="@daily",
 )
-# [START howto_operator_datetime_branch_next_day]
+# [START howto_branch_datetime_operator_next_day]
 dummy_task_1 = DummyOperator(task_id='date_in_range', dag=dag)
 dummy_task_2 = DummyOperator(task_id='date_outside_range', dag=dag)
 
-cond2 = DateTimeBranchOperator(
+cond2 = BranchDateTimeOperator(
     task_id='datetime_branch',
     follow_task_ids_if_true=['date_in_range'],
     follow_task_ids_if_false=['date_outside_range'],
@@ -80,4 +80,4 @@ cond2 = DateTimeBranchOperator(
 # Since target_lower happens after target_upper, target_upper will be moved to the following day
 # Run dummy_task_1 if cond2 executes between 15:00:00, and 00:00:00 of the following day
 cond2 >> [dummy_task_1, dummy_task_2]
-# [END howto_operator_datetime_branch_next_day]
+# [END howto_branch_datetime_operator_next_day]
