+++ docs/apache-airflow/howto/operator/datetime.rst
@@ -17,23 +17,23 @@
 
 
 
-.. _howto/operator:DatetimeBranch:
+.. _howto/operator:BranchDateTimeOperator:
 
-DatetimeBranchOperator
+BranchDateTimeOperator
 ======================
 
-Use the :class:`~airflow.operators.datetime_branch.DatetimeBranchOperator` to branch into one of two execution paths depending on whether the date and/or time of execution falls into the range given by two target arguments.
+Use the :class:`~airflow.operators.datetime.BranchDateTimeOperator` to branch into one of two execution paths depending on whether the date and/or time of execution falls into the range given by two target arguments.
 
-.. exampleinclude:: /../../airflow/example_dags/example_datetime_branch_operator.py
+.. exampleinclude:: /../../airflow/example_dags/example_branch_datetime_operator.py
     :language: python
-    :start-after: [START howto_operator_datetime_branch]
-    :end-before: [END howto_operator_datetime_branch]
+    :start-after: [START howto_branch_datetime_operator]
+    :end-before: [END howto_branch_datetime_operator]
 
 The target parameters, ``target_upper`` and ``target_lower``, can receive a ``datetime.datetime``, a ``datetime.time``, or ``None``. When a ``datetime.time`` object is used, it will be combined with the current date in order to allow comparisons with it. In the event that ``target_upper`` is set to a ``datetime.time`` that occurs before the given ``target_lower``, a day will be added to ``target_upper``. This is done to allow for time periods that span over two dates.
 
-.. exampleinclude:: /../../airflow/example_dags/example_datetime_branch_operator.py
+.. exampleinclude:: /../../airflow/example_dags/example_branch_datetime_operator.py
     :language: python
-    :start-after: [START howto_operator_datetime_branch_next_day]
-    :end-before: [END howto_operator_datetime_branch_next_day]
+    :start-after: [START howto_branch_datetime_operator_next_day]
+    :end-before: [END howto_branch_datetime_operator_next_day]
 
 If a target parameter is set to ``None``, the operator will perform a unilateral comparison using only the non-``None`` target. Setting both ``target_upper`` and ``target_lower`` to ``None`` will raise an exception.
